D23595: [PowerPC] INTRNG preparation: Rename our pic interface
To allow for porting to the INTRNG framework, we need to be able to include the MI pic interface. Since we have an interface name conflict with it, we first have to change the name of ours.

Rename the powerpc pic interface to oldpic so they can coexist, and to help with spotting where code needs to be updated.
Update all pic interface calls in the PPC code.

Should be NFC.